requirements.txt
APScheduler==3.10.4
Flask==3.0.2
flask_restx==1.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
